Move erase to plot and update erased files
